# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Food & Handmade Products Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/profile:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    put:
      operationId: updateUserProfile
      tags: [users]
      summary: Update user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/addresses:
    get:
      operationId: getAddresses
      tags: [users]
      summary: Get user addresses
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of addresses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Address"
    post:
      operationId: addAddress
      tags: [users]
      summary: Add new address
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressBody"
      responses:
        "201":
          description: Address added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
  /users/addresses/{id}:
    put:
      operationId: updateAddress
      tags: [users]
      summary: Update address
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressBody"
      responses:
        "200":
          description: Address updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
    delete:
      operationId: deleteAddress
      tags: [users]
      summary: Delete address
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Address deleted

  # PRODUCTS
...
  /sellers/dashboard:
    get:
      operationId: getSellerDashboard
      tags: [sellers]
      summary: Get seller dashboard stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellerDashboard"
  /sellers/profile:
    get:
      operationId: getSellerProfile
      tags: [sellers]
      summary: Get current seller profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Seller profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellerProfile"
        "404":
          description: Not a seller
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createSellerProfile
      tags: [sellers]
      summary: Create seller profile (apply to become seller)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SellerProfileBody"
      responses:
        "201":
          description: Seller profile created (pending approval)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellerProfile"
    put:
      operationId: updateSellerProfile
      tags: [sellers]
      summary: Update seller profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SellerProfileBody"
      responses:
        "200":
          description: Seller profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellerProfile"
  /sellers/orders:
    get:
      operationId: getSellerOrders
      tags: [sellers]
      summary: Get seller's orders
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Seller order list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
  /sellers/sales-history:
    get:
      operationId: getSellerSalesHistory
      tags: [sellers]
      summary: Get seller sales history and stats
      security:
        - bearerAuth: []
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Sales history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesHistory"
  /sellers/approve/{userId}:
    put:
      operationId: approveSeller
      tags: [sellers]
      summary: Approve seller (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Seller approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellerProfile"
  /sellers/{id}:
    get:
      operationId: getPublicSellerProfile
      tags: [sellers]
      summary: Get public seller profile by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Seller profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellerProfile"

  # CART
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error

    # Auth
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
        sellerApproved:
          type: boolean
        addresses:
          type: array
          items:
            $ref: "#/components/schemas/Address"
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - phone
        - role
        - sellerApproved
        - addresses
        - createdAt
    UpdateProfileBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
    Address:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        street:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        phone:
          type: string
        isDefault:
          type: boolean
      required:
        - id
        - userId
        - name
        - street
        - city
        - state
        - zipCode
        - phone
        - isDefault
    AddressBody:
      type: object
      properties:
        name:
          type: string
        street:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        phone:
          type: string
        isDefault:
          type: boolean
      required:
        - name
        - street
        - city
        - state
        - zipCode
        - phone

    # Products
...
    OrderItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        price:
          type: number
        subtotal:
          type: number
        unit:
          type: string
      required:
        - productId
        - productName
        - quantity
        - price
        - subtotal
        - unit
    DeliveryAddress:
      type: object
      properties:
        name:
          type: string
        street:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        phone:
          type: string
      required:
        - name
        - street
        - city
        - state
        - zipCode
        - phone
    Order:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: string
        buyerId:
          type: integer
        sellerId:
          type: integer
        buyerName:
          type: string
        sellerName:
          type: string
        shopName:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        totalPrice:
          type: number
        deliveryAddress:
          $ref: "#/components/schemas/DeliveryAddress"
        orderMode:
          type: string
          enum: [ecommerce, whatsapp]
        status:
          type: string
          enum: [pending, confirmed, shipped, delivered, cancelled]
        paymentMethod:
          type: string
        paymentStatus:
          type: string
          enum: [pending, completed, failed]
        sellerNotes:
          type: string
        trackingNumber:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - orderId
        - buyerId
        - sellerId
        - buyerName
        - sellerName
        - shopName
        - items
        - totalPrice
        - deliveryAddress
        - orderMode
        - status
        - paymentStatus
        - createdAt
        - updatedAt
    OrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - orders
        - total
        - page
        - totalPages
...
    SellerProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        shopName:
          type: string
        phone:
          type: string
        address:
          type: string
        orderingMode:
          type: string
          enum: [ecommerce, whatsapp]
        whatsappNumber:
          type: string
        paymentMethods:
          type: array
          items:
            type: string
        verified:
          type: boolean
        userName:
          type: string
        userEmail:
          type: string
        totalProducts:
          type: integer
        totalOrders:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - userId
        - shopName
        - phone
        - address
        - orderingMode
        - paymentMethods
        - verified
        - userName
        - userEmail
        - createdAt
    SellerProfileBody:
      type: object
      properties:
        shopName:
          type: string
        phone:
          type: string
        address:
          type: string
        orderingMode:
          type: string
          enum: [ecommerce, whatsapp]
        whatsappNumber:
          type: string
        paymentMethods:
          type: array
          items:
            type: string
      required:
        - shopName
        - phone
        - address
        - orderingMode
    SellerDashboard:
      type: object
      properties:
        totalProducts:
          type: integer
        totalOrders:
          type: integer
        pendingOrders:
          type: integer
        totalRevenue:
          type: number
        recentOrders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        orderingMode:
          type: string
      required:
        - totalProducts
        - totalOrders
        - pendingOrders
        - totalRevenue
        - recentOrders
        - orderingMode
    SalesHistory:
      type: object
      properties:
        totalSales:
          type: number
        totalOrders:
          type: integer
        avgOrderValue:
          type: number
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
      required:
        - totalSales
        - totalOrders
        - avgOrderValue
        - orders

    # Cart
...